requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==2.2.8
django-admin-rangefilter==0.5.1
django-crispy-forms==1.8.1
django-taggit==1.2.0
django-tinymce==2.8.0
djangorestframework==3.11.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
Pillow==6.2.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.7